% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hello_underworld}
\alias{.hello_underworld}
\title{An Example Internal Function}
\usage{
.hello_underworld(x, ...)
}
\arguments{
\item{x}{a single character string}

\item{...}{other arguments passed to \link[stats:as.formula]{stats::as.formula}}
}
\value{
a \code{formula}
}
\description{
Does not do anything, just serves as an example of an internal function within a project.
It is internal because it starts with a dot.
}
\details{
Arguments are checked using \link[assertthat:assert_that]{assertthat::assert_that}.
}
