% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hello_world}
\alias{hello_world}
\title{An Example of Exported Function}
\usage{
hello_world(x, y, z, verbose = TRUE, prefix = "a")
}
\arguments{
\item{x}{a \code{matrix} object with at least 10 rows (see \strong{Details} section)}

\item{y}{a single positive number}

\item{z}{a character vector with no empty elements}

\item{verbose}{a logical that indicates if log messages
should be output}

\item{prefix}{a single character string that specifies the prefix
to use.}
}
\value{
an \code{ExpressionSet} object
}
\description{
Does not do anything, just serves as an example of an internal function within a project.
It is exported because it does \strong{not} start with a dot.
}
\details{
Arguments are checked using \link[assertthat:assert_that]{assertthat::assert_that}.
}
\examples{
x <- matrix(1:20, ncol = 2)
hello_world(x, 2, "abcd")

}
